maj du makefile

+++ srcs/parsing.c
@@ -6,7 +6,7 @@
 /*   By: sgabsi <sgabsi@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/22 15:12:07 by sgabsi            #+#    #+#             */
-/*   Updated: 2024/03/14 13:44:42 by sgabsi           ###   ########.fr       */
+/*   Updated: 2024/03/14 13:54:23 by sgabsi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,7 +81,7 @@ static int	check_overflow(const char *str, int *val)
 			return (1);
 		i++;
 	}
-	*val = num;
+	*val = (int) num;
 	if (sign < 0)
 		(*val) *= sign;
 	return (0);
